Bug fixes (only color left to be calculated)
